// src/component/appcontainer/alldefaultprops/Enums.ts
enum DropableControlElementEnums {
    SiteName = "_Site",
    RoomName = "_Room",
    FloorName = "_Floor",
    LocationName = "_Location",
    DeviceName = "_Device"
}

enum deviceModelTabs {
    Search = "Search",
    Property = "Property",
    Related = "Related",
    Result = "Result"
}

export { DropableControlElementEnums, deviceModelTabs }